backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "My Project [Travel Journal] Backend"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "alembic>=1.18.4",
    "bcrypt>=5.0.0",
    "celery>=5.6.2",
    "fastapi>=0.135.1",
    "httpx>=0.28.1",
    "passlib[bcrypt]>=1.7.4",
    "psycopg2-binary>=2.9.11",
    "pydantic-settings>=2.13.1",
    "pydantic[email]>=2.12.5",
    "python-jose[cryptography]>=3.5.0",
    "python-multipart>=0.0.22",
    "redis>=7.3.0",
    "sqlalchemy>=2.0.48",
    "uvicorn>=0.41.0",
]
